% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add_prefixes}
\alias{add_prefixes}
\title{Add SPARQL prefixes to a query.}
\usage{
add_prefixes(query)
}
\arguments{
\item{query}{A character string representing the SPARQL query to which prefixes need to be added.}
}
\value{
A character string, which is the input query with SPARQL prefixes added.
}
\description{
The `add_prefixes` function is designed to add SPARQL prefixes to a given query.
It takes a query as input and prefixes it with the specified namespace prefixes.
}
\details{
The function utilizes an inner function `format.namespace` to format each namespace prefix
and its corresponding URI as a SPARQL PREFIX statement. The prefixes are defined in a
global variable named `prefixes`. The format of the global variable should be a named list
where each name corresponds to a prefix, and the corresponding value is the URI.
}
\examples{
# Define namespace prefixes
prefixes <- list(
  ex = "http://example.org/",
  foaf = "http://xmlns.com/foaf/0.1/",
  rdf = "http://www.w3.org/1999/02/22-rdf-syntax-ns#"
)

# Example query
query <- "SELECT ?person WHERE { ?person foaf:name ?name . }"

# Add prefixes to the query
result_query <- add_prefixes(query)

# Print the result
cat(result_query)

}
\author{
<Author 1>
}
