% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fill.query}
\alias{fill.query}
\title{Fill in placeholders in a SPARQL query based on provided parameters.}
\usage{
fill.query(query, taxid = NULL, sciName = NULL)
}
\arguments{
\item{query}{A character string representing the SPARQL query with placeholders.}

\item{taxid}{An optional parameter representing the Taxonomy ID to be used in the query.}

\item{sciName}{An optional parameter representing the scientific name to be used in the query.}
}
\value{
A character string representing the filled-in SPARQL query with specified values.
}
\description{
The `fill.query` function replaces placeholders in a SPARQL query with specified
values based on taxonomy ID or scientific name. It utilizes the `get.sparql.values`,
`get.named.graphs`, `query.by.name`, and `query.by.taxid` functions.
}
\details{
The function takes a SPARQL query with placeholders and optional parameters for Taxonomy ID
and scientific name. It uses the `get.sparql.values` function to generate VALUES clauses for
the provided parameters and replaces the placeholders in the query using `sprintf`. The
function also includes information from `get.named.graphs`, `query.by.name`, and `query.by.taxid`
in the resulting query.
}
\examples{
# Example for filling in placeholders in a SPARQL query
sparql_query_template <- "SELECT * WHERE { \%s }"
filled_query <- fill.query(sparql_query_template, taxid = c("GBIF:2130185", "NCBI:55786"))
cat(filled_query)

}
\author{
<Author 1>
}
