% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{format.result.df}
\alias{format.result.df}
\title{Format and process a data frame of results.}
\usage{
\method{format}{result.df}(df, iri.cols, with.matchId = TRUE, with.reference = TRUE)
}
\arguments{
\item{df}{A data frame containing query results.}

\item{iri.cols}{A vector of character strings specifying the columns containing IRIs to be transformed.}

\item{with.matchId}{Logical, indicating whether to group matching taxids into a list.}

\item{with.reference}{Logical, indicating whether to group bibliographic references into a list.}
}
\value{
A modified data frame with transformed IRIs, grouped taxids, and deduplicated rows.
}
\description{
The `format.result.df` function transforms IRIs in a data frame to their short, prefixed form
using the `to.short.iris` function. It also groups matching taxids into a list if specified,
and then deduplicates the data frame.
}
\details{
The function transforms specified columns containing IRIs to their short, prefixed form using
the `to.short.iris` function. It also provides an option to group matching taxids into a list
if the `with.matchId` parameter is set to TRUE. The resulting data frame is then deduplicated.
}
\examples{
# Example for formatting and processing a data frame of results
result_df <- data.frame(matchId = c(1, 1, 2, 3),
                        iri_col = c("http://example.org/1", "http://example.org/2", "http://example.org/3", "http://example.org/4"))
formatted_df <- format.result.df(result_df, iri.cols = "iri_col", with.matchId = TRUE, with.reference=FALSE)
print(formatted_df)

}
\author{
<Author 1>
}
