% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get.named.graphs}
\alias{get.named.graphs}
\title{Generate a SPARQL FROM NAMED clause for named graphs.}
\usage{
get.named.graphs()
}
\value{
A character string representing the SPARQL FROM NAMED clause with specified named graphs.
}
\description{
The `get.named.graphs` function generates a SPARQL FROM NAMED clause with specified named graphs.
The resulting SPARQL snippet can be used to specify named graphs for a query.
}
\details{
The function returns a static SPARQL snippet with FROM NAMED clauses for three example named graphs.
}
\examples{
# Example for generating a SPARQL FROM NAMED clause
named_graphs <- get.named.graphs()
cat(named_graphs)

}
\author{
<Author 1>
}
