% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get.source.metadata}
\alias{get.source.metadata}
\title{Retrieve metadata for a specified data source from a specific SPARQL endpoint.}
\usage{
get.source.metadata(source, endpoint = NULL)
}
\arguments{
\item{source}{A character string representing the shortened IRI of the data source for which metadata is to be retrieved.}

\item{endpoint}{A character string representing the SPARQL endpoint to send the query to.}
}
\value{
A data frame containing metadata for the specified data source.
}
\description{
The `get.source.metadata` function constructs and executes a SPARQL query to retrieve metadata
for a specified data source. It utilizes the `get.sparql.values`, `send.sparql`, and `format.result.df`
functions.
}
\details{
The function constructs a SPARQL query template, fills in placeholders based on provided parameters
using the `get.sparql.values` function, sends the query to the specified SPARQL endpoint using `send.sparql`,
and formats the result data frame using `format.result.df`. The resulting data frame contains metadata
information such as field and value for the specified data source.
}
\examples{
# Example for retrieving metadata for a specific data source from a SPARQL endpoint
source_metadata <- get.source.metadata(source = "gratin:nematrait", endpoint = "http://129.88.204.79:7200/repositories/gratin")
print(source_metadata)

}
\author{
<Author 1>
}
