% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get.sources}
\alias{get.sources}
\title{Retrieve a list of data sources from a specific SPARQL endpoint.}
\usage{
get.sources(endpoint = NULL)
}
\arguments{
\item{endpoint}{A character string representing the SPARQL endpoint to send the query to.}
}
\value{
A data frame containing a list of data sources available at the specified SPARQL endpoint.
}
\description{
The `get.sources` function constructs and executes a SPARQL query to retrieve a list of data sources
from a specified SPARQL endpoint. It utilizes the `request`, `req_retry`, `req_perform`, `resp_check_status`,
`read.table`, and `format.result.df` functions.
}
\details{
The function constructs a SPARQL query to obtain the list of data sources, sends the query to the specified
SPARQL endpoint using `request`, `req_retry`, and `req_perform`, checks the response status with `resp_check_status`,
reads the response body with `read.table`, and formats the result data frame using `format.result.df`. The resulting
data frame contains information about data sources, specifically the source shortened IRIs.
}
\examples{
# Example for retrieving a list of data sources from a SPARQL endpoint
sources_list <- get.sources(endpoint = "http://129.88.204.79:7200/repositories/gratin")
print(sources_list)

}
\author{
<Author 1>
}
