% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get.sparql.values}
\alias{get.sparql.values}
\title{Generate a SPARQL VALUES clause for a variable with specified values.}
\usage{
get.sparql.values(var.name, var.values, is.iri = TRUE)
}
\arguments{
\item{var.name}{A character string representing the variable name for which values are specified.}

\item{var.values}{A vector of values for the specified variable.}

\item{is.iri}{Logical, indicating whether the values are IRIs. If FALSE, the values are wrapped
in double quotes.}
}
\value{
A character string representing the SPARQL VALUES clause for the specified variable and values.
}
\description{
The `get.sparql.values` function generates a SPARQL VALUES clause for a specified variable
with associated values. The resulting SPARQL snippet can be used to filter query results based on
the specified variable values.
}
\details{
If `is.iri` is set to FALSE, the function wraps each value in double quotes. The resulting
VALUES clause is then constructed and returned as a character string.
}
\examples{
# Example for generating a SPARQL VALUES clause
var_name <- "country"
var_values <- c("USA", "Canada", "UK")
sparql_values <- get.sparql.values(var_name, var_values, is.iri = FALSE)
cat(sparql_values)

}
\author{
<Author 1>
}
