% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{query.by.name}
\alias{query.by.name}
\title{Generate a static SPARQL snippet for retrieving information by taxon scientific name.}
\usage{
query.by.name()
}
\value{
A character string representing the static SPARQL snippet for retrieving information by taxon scientific name.
}
\description{
The `query.by.name` function generates a static SPARQL snippet for retrieving information based on
a taxon scientific name.
}
\examples{
# Example for generating a SPARQL query by taxon scientific name
sparql_query <- query.by.name()
cat(sparql_query)

}
\author{
<Author 1>
}
