% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{send.sparql}
\alias{send.sparql}
\title{Send a SPARQL query to an endpoint and retrieve results.}
\usage{
send.sparql(query, endpoint, verbose = FALSE)
}
\arguments{
\item{query}{A character string representing the SPARQL query to be sent.}

\item{endpoint}{A character string representing the SPARQL endpoint to send the query to.}

\item{verbose}{Logical, indicating whether to print the SPARQL query before sending.}
}
\value{
A tibble containing the results of the SPARQL query.
}
\description{
The `send.sparql` function sends a SPARQL query to a specified endpoint and retrieves the results.
The function also supports adding prefixes to the query using the `add_prefixes` function.
}
\details{
The function sends a GET request to the specified SPARQL endpoint with the provided query.
It expects the response in the SPARQL Results JSON format and parses the results into a tibble.
The function also checks the response type and raises an error if it's not in the expected format.
}
\examples{
# Example for sending a SPARQL query to an endpoint
query <- "SELECT ?subject ?predicate ?object WHERE { ?subject ?predicate ?object } LIMIT 10"
endpoint <- "http://example.com/sparql"
result <- send.sparql(query, endpoint, verbose = TRUE)
print(result)

}
\author{
<Author 1>
}
