% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{source.tracking}
\alias{source.tracking}
\title{Generate a static SPARQL snippet for tracking the source of information about an organism}
\usage{
source.tracking()
}
\value{
A character string representing the static SPARQL snippet for tracking the source of information about an organism.
}
\description{
This function returns a SPARQL query string that retrieves the source of information about the focal organism,
and includes a binding to check if the information has been inferred.
}
\examples{
query <- source.tracking()
cat(query)

}
\author{
<Author 1>
}
