% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{to.short.iri}
\alias{to.short.iri}
\title{Shorten a full IRI using specified namespaces.}
\usage{
to.short.iri(full.iri, namespaces = prefixes)
}
\arguments{
\item{full.iri}{A character string representing the full IRI to be shortened.}

\item{namespaces}{A named list containing namespace prefixes and their corresponding URIs.}
}
\value{
A character string representing the shortened IRI with the associated prefix.
}
\description{
The `to.short.iri` function shortens a full Internationalized Resource Identifier (IRI)
using specified namespaces. It attempts to find the longest matching namespace and replaces
it with its associated prefix.
}
\details{
The function iterates through the provided namespaces to find the longest matching namespace
for the given full IRI. If a matching namespace is found, it replaces the matching part with
the associated prefix and returns the shortened IRI. If no matching namespace is found, the
original full IRI is returned unchanged.
}
\examples{
# Example for shortening a full IRI using namespaces
full_iri <- "http://example.org/resource"
namespaces <- list(ex = "http://example.org/", foaf = "http://xmlns.com/foaf/0.1/")
shortened_iri <- to.short.iri(full_iri, namespaces)
cat(shortened_iri)

}
\author{
<Author 1>
}
