% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{to.short.iris}
\alias{to.short.iris}
\title{Shorten multiple full IRIs using specified namespaces.}
\usage{
to.short.iris(iris)
}
\arguments{
\item{iris}{A vector of character strings representing the full IRIs to be shortened.}
}
\value{
A vector of character strings representing the shortened IRIs.
}
\description{
The `to.short.iris` function shortens a vector of full Internationalized Resource Identifiers (IRIs)
using the `to.short.iri` function. It applies the `to.short.iri` function to each element of the vector.
}
\details{
The function applies the `to.short.iri` function to each element of the input vector of full IRIs.
It returns a vector of shortened IRIs with associated prefixes based on the specified namespaces.
}
\examples{
# Example for shortening multiple full IRIs using namespaces
full_iris <- c("http://example.org/resource1", "http://example.org/resource2")
shortened_iris <- to.short.iris(full_iris)
cat(shortened_iris)

}
\author{
<Author 1>
}
